rename pinnable
